PBL/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.0
chromadb>=0.4.0
torch>=2.0.0
transformers>=4.30.0
requests>=2.28.0
python-docx>=0.8.11
PyPDF2>=3.0.0
nbformat>=5.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0